Split RankableOptions and polls.py into smaller modules
Two of the largest files in the repo were nearly 2000 and 1100 lines.
Both had clean, dependency-free seams that could be lifted out without
restructuring callers.

components/RankableOptions.tsx (2002 → 1711 lines):
- components/rankable/tiers.ts: pairKey, computeTierIndices, tiersFromList,
  computeDropTarget, getTierRange, tierRanks. Re-exported from
  RankableOptions.tsx so existing test imports still resolve.
- components/rankable/visuals.tsx: LinkIcon, GripLines, DragHandleVisual,
  TierCardRows, LinkCircle, plus the three layout constants.

server/routers/polls.py (1063 → 410 lines):
- server/services/polls.py: every non-route helper — _SELECT_POLL_FULL,
  _fetch_poll_full, _attach_wrapper_fields, _check_auto_close,
  _finalize_suggestion_options, _finalize_time_slots, _row_to_poll,
  _row_to_vote, _enforce_suggestion_phase_timing, _submit_vote_to_poll,
  _edit_vote_on_poll, _compute_results, _json_or_none.
- routers/multipolls.py now imports these from services.polls instead of
  reaching into routers.polls — eliminates the cross-router back-reference.

Verified: tsc clean, 160 vitest tests pass, 152 pytest unit tests pass,
FastAPI app boots with all 30 routes registered.
