+++ components/RankableOptions.tsx
@@ -5,6 +5,33 @@ import { flushSync } from 'react-dom';
 import { ClientOnlyDragDrop } from './ClientOnly';
 import type { OptionsMetadata } from '@/lib/types';
 import OptionLabel, { isLocationEntry } from './OptionLabel';
+import {
+  pairKey,
+  computeTierIndices,
+  tiersFromList,
+  computeDropTarget,
+  getTierRange,
+  tierRanks,
+} from './rankable/tiers';
+import {
+  LINK_CIRCLE_SIZE,
+  LINK_CONTOUR_FILTER,
+  LinkIcon,
+  DragHandleVisual,
+  TierCardRows,
+  LinkCircle,
+} from './rankable/visuals';
+
+// Re-export pure helpers so existing test imports
+// (`@/components/RankableOptions`) keep working.
+export {
+  pairKey,
+  computeTierIndices,
+  tiersFromList,
+  computeDropTarget,
+  getTierRange,
+  tierRanks,
+};
 
 interface RankableOption {
   id: string;
@@ -33,325 +60,6 @@ interface RankableOptionsProps {
   disableGrouping?: boolean;
 }
 
-/**
- * Canonical key for a pair of item ids, order-independent. Used so that
- * linked-pair membership survives when a linked-group is reordered end-to-end.
- */
-export const pairKey = (a: string, b: string): string => (a < b ? `${a}|${b}` : `${b}|${a}`);
-
-/**
- * Compute tier boundaries from a flat list + set of linked pairs.
- * Returns an array of tiers, where each tier is a list of indices into the
- * main list that belong to that tier.
- */
-export function computeTierIndices(
-  list: readonly { id: string }[],
-  linkedPairs: ReadonlySet<string>,
-): number[][] {
-  if (list.length === 0) return [];
-  const tiers: number[][] = [[0]];
-  for (let i = 1; i < list.length; i++) {
-    const prev = list[i - 1];
-    const curr = list[i];
-    if (linkedPairs.has(pairKey(prev.id, curr.id))) {
-      tiers[tiers.length - 1].push(i);
-    } else {
-      tiers.push([i]);
-    }
-  }
-  return tiers;
-}
-
-/** Convert tier index ranges into the `[["A"], ["B","C"], ...]` shape. */
-export function tiersFromList(
-  list: readonly { id: string; text: string }[],
-  linkedPairs: ReadonlySet<string>,
-): string[][] {
-  return computeTierIndices(list, linkedPairs).map(tier =>
-    tier.map(i => list[i].text),
-  );
-}
-
-/**
- * Small round button that toggles whether two adjacent items are tied.
- * Uses the same chain-link glyph in both states — just blue when active
- * (linked) and muted gray when inactive (unlinked).
- *
- * Centered on the left edge of the cards column so the circle overlaps the
- * corners of the two items it sits between.
- */
-const LINK_CIRCLE_SIZE = 24;
-const LINK_ICON_SIZE = 15;
-const LINK_CONTOUR_FILTER = 'drop-shadow(0 0 3px var(--background)) drop-shadow(0 0 3px var(--background)) drop-shadow(0 0 3px var(--background)) drop-shadow(0 0 3px var(--background)) drop-shadow(0 0 3px var(--background)) drop-shadow(0 0 3px var(--background))';
-
-/** The chain-link SVG glyph reused in link circles and drag preview links. */
-function LinkIcon({ size = LINK_ICON_SIZE }: { size?: number }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.25" strokeLinecap="round" strokeLinejoin="round">
-      <path d="M10 13a5 5 0 0 0 7.54.54l3-3a5 5 0 0 0-7.07-7.07l-1.72 1.71" />
-      <path d="M14 11a5 5 0 0 0-7.54-.54l-3 3a5 5 0 0 0 7.07 7.07l1.71-1.71" />
-    </svg>
-  );
-}
-
-/** The up/down arrows + grip lines used in both live and preview drag handles. */
-function DragHandleVisual({ dimUp = false, variant = 'main' }: { dimUp?: boolean; variant?: 'main' | 'noPreference' }) {
-  if (variant === 'noPreference') {
-    return (
-      <div className="flex flex-col items-center justify-center gap-0">
-        <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round" className="text-green-500 dark:text-green-400">
-          <line x1="12" y1="5" x2="12" y2="19" /><line x1="5" y1="12" x2="19" y2="12" />
-        </svg>
-        <GripLines />
-        <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round" className="text-green-500 dark:text-green-400">
-          <line x1="12" y1="5" x2="12" y2="19" /><line x1="5" y1="12" x2="19" y2="12" />
-        </svg>
-      </div>
-    );
-  }
-  return (
-    <div className="flex flex-col items-center justify-center gap-0">
-      <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round" className={dimUp ? 'text-gray-300 dark:text-gray-600' : 'text-gray-400 dark:text-gray-500'}>
-        <polyline points="18 15 12 9 6 15" />
-      </svg>
-      <GripLines />
-      <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round" className="text-gray-400 dark:text-gray-500">
-        <polyline points="6 9 12 15 18 9" />
-      </svg>
-    </div>
-  );
-}
-
-function GripLines() {
-  return (
-    <div className="flex flex-col items-center justify-center my-0.5">
-      <div className="w-3.5 h-0.5 bg-gray-300 dark:bg-gray-600 mb-0.5" />
-      <div className="w-3.5 h-0.5 bg-gray-300 dark:bg-gray-600 mb-0.5" />
-      <div className="w-3.5 h-0.5 bg-gray-300 dark:bg-gray-600" />
-    </div>
-  );
-}
-
-/**
- * Renders the interior of a tier card: rows of option content with divider
- * lines between them. Used by both the live ballot and the floating drag
- * preview so the visual structure stays in sync.
- */
-function TierCardRows({
-  items,
-  stepSize,
-  gapSize: innerGap,
-  itemHeight: rowHeight,
-  dividerClassName,
-  dividerInset = { left: '12px', right: '12px' },
-  renderContent,
-  rowProps,
-}: {
-  items: { id: string; text: string }[];
-  stepSize: number;
-  gapSize: number;
-  itemHeight: number;
-  dividerClassName: string;
-  dividerInset?: { left: string; right: string };
-  renderContent: (item: { id: string; text: string }, index: number) => React.ReactNode;
-  rowProps?: (item: { id: string; text: string }, rowIdx: number) => Record<string, unknown>;
-}) {
-  return (
-    <>
-      {items.map((item, rowIdx) => {
-        const rowTop = rowIdx * stepSize;
-        const extra = rowProps?.(item, rowIdx) ?? {};
-        return (
-          <React.Fragment key={item.id}>
-            <div
-              className="absolute left-0 right-0 p-3"
-              style={{ top: `${rowTop}px`, height: `${rowHeight}px` }}
-              {...extra}
-            >
-              <div className="flex items-center justify-between h-full">
-                {renderContent(item, rowIdx)}
-              </div>
-            </div>
-            {rowIdx < items.length - 1 && (
-              <div
-                className={`absolute pointer-events-none border-t ${dividerClassName}`}
-                style={{
-                  top: `${rowTop + rowHeight + innerGap / 2}px`,
-                  left: dividerInset.left,
-                  right: dividerInset.right,
-                  transform: 'translateY(-0.5px)',
-                }}
-              />
-            )}
-          </React.Fragment>
-        );
-      })}
-    </>
-  );
-}
-
-function LinkCircle({
-  entry,
-  disabled,
-  onToggle,
-  translateY,
-}: {
-  entry: { topCenter: number; linked: boolean; idA: string; idB: string };
-  disabled: boolean;
-  onToggle: (idA: string, idB: string) => void;
-  translateY?: number;
-}) {
-  const { topCenter, linked, idA, idB } = entry;
-  return (
-    <button
-      type="button"
-      disabled={disabled}
-      onClick={(e) => {
-        e.preventDefault();
-        e.stopPropagation();
-        onToggle(idA, idB);
-      }}
-      className={`absolute flex items-center justify-center transition-colors ${
-        disabled
-          ? 'cursor-not-allowed text-gray-400 dark:text-gray-600'
-          : linked
-            ? 'cursor-pointer text-blue-600 hover:text-blue-700 dark:text-blue-400 dark:hover:text-blue-300'
-            : 'cursor-pointer text-gray-400 hover:text-blue-500 dark:text-gray-500 dark:hover:text-blue-400'
-      }`}
-      style={{
-        left: '50%',
-        top: `${topCenter - LINK_CIRCLE_SIZE / 2}px`,
-        width: `${LINK_CIRCLE_SIZE}px`,
-        height: `${LINK_CIRCLE_SIZE}px`,
-        zIndex: 3,
-        transform: `translateX(-50%)${translateY ? ` translateY(${translateY}px)` : ''}`,
-        // Background-colored contour around the icon for contrast against
-        // the card surface it overlaps. Stacked drop-shadows thicken the
-        // halo because a single pass is too faint.
-        filter: LINK_CONTOUR_FILTER,
-      }}
-      aria-label={linked ? 'Break tied ranking' : 'Tie these rankings together'}
-      title={linked ? 'Break tied ranking' : 'Tie these rankings together'}
-    >
-      <LinkIcon />
-    </button>
-  );
-}
-
-/**
- * Pure, testable drop-target computation for main-list drags. Given the
- * dragged tier's visual center (from cursor + grab offset) and the list
- * layout, returns the original-list index where the tier should land.
- *
- * Algorithm: for every valid insertion point (between non-dragged units
- * or at the edges), compute where the tier's center would be in the
- * resulting layout. Return the insertion point whose layout center is
- * closest to the tier's actual visual center. This naturally gives
- * symmetric thresholds (midpoint between two adjacent layout centers)
- * and treats linked groups as atomic.
- */
-export function computeDropTarget(
-  itemIds: readonly string[],
-  linkedPairs: ReadonlySet<string>,
-  tierStart: number,
-  tierSize: number,
-  tierVisualCenter: number,
-  itemHeight: number,
-  groupedGapSize: number,
-  totalItemHeight: number,
-): number {
-  const tEnd = tierStart + tierSize - 1;
-  const tierH = tierSize > 1
-    ? tierSize * itemHeight + (tierSize - 1) * groupedGapSize
-    : itemHeight;
-
-  // Build non-dragged unit groups
-  const allTiers = computeTierIndices(
-    itemIds.map(id => ({ id })),
-    linkedPairs,
-  );
-
-  // Collect valid target indices (boundaries between non-dragged units)
-  const validTargets: number[] = [];
-  for (const tier of allTiers) {
-    const first = tier[0];
-    const last = tier[tier.length - 1];
-    if (first >= tierStart && last <= tEnd) continue; // skip dragged tier
-    validTargets.push(first); // insert-before-this-unit position
-  }
-  // After the last non-dragged unit
-  const lastNonDragged = allTiers.filter(
-    t => !(t[0] >= tierStart && t[t.length - 1] <= tEnd),
-  );
-  if (lastNonDragged.length > 0) {
-    const last = lastNonDragged[lastNonDragged.length - 1];
-    validTargets.push(last[last.length - 1] + 1);
-  }
-
-  if (validTargets.length === 0) return tierStart;
-
-  // For each valid target, compute the tier's "natural center" in the
-  // resulting layout. Each item takes one totalItemHeight slot. The tier
-  // occupies tierSize slots starting at "slot = number of non-dragged
-  // items before the insertion point".
-  let bestTarget = validTargets[0];
-  let bestDist = Infinity;
-
-  for (const target of validTargets) {
-    // Count non-dragged items whose original index < target
-    let itemsBefore = 0;
-    for (let i = 0; i < itemIds.length; i++) {
-      if (i >= tierStart && i <= tEnd) continue;
-      if (i < target) itemsBefore++;
-    }
-    const tierNaturalCenter = itemsBefore * totalItemHeight + tierH / 2;
-    const dist = Math.abs(tierVisualCenter - tierNaturalCenter);
-    if (dist < bestDist) {
-      bestDist = dist;
-      bestTarget = target;
-    }
-  }
-
-  return bestTarget;
-}
-
-/**
- * Return `[tierStart, tierSize]` for the tier containing `index` in the
- * given list + links. A tier is a maximal run of consecutive items connected
- * by linked pairs. An untied item returns `[index, 1]`.
- */
-export function getTierRange(
-  list: readonly { id: string }[],
-  linkedPairs: ReadonlySet<string>,
-  index: number,
-): [number, number] {
-  if (index < 0 || index >= list.length) return [index, 1];
-  // Walk backward while the previous pair is linked
-  let start = index;
-  while (start > 0 && linkedPairs.has(pairKey(list[start - 1].id, list[start].id))) {
-    start--;
-  }
-  // Walk forward while the next pair is linked
-  let end = index;
-  while (end < list.length - 1 && linkedPairs.has(pairKey(list[end].id, list[end + 1].id))) {
-    end++;
-  }
-  return [start, end - start + 1];
-}
-
-/**
- * Compute standard-competition ranks for a tiered ballot (1, 2, 2, 4).
- * Returns one integer per tier: the display rank for that tier.
- */
-export function tierRanks(tiers: readonly (readonly unknown[])[]): number[] {
-  const ranks: number[] = [];
-  let pos = 0;
-  for (const tier of tiers) {
-    ranks.push(pos + 1);
-    pos += tier.length;
-  }
-  return ranks;
-}
 
 export default function RankableOptions({ options, onRankingChange, disabled = false, storageKey, initialRanking, initialTiers, optionsMetadata, renderOption, preserveOrder = false, disableGrouping = false, newOptions }: RankableOptionsProps) {
